VZ-6318 Support custom Jaeger secret in verrazzano-install namespace (#3744)
To support external OpenSearch as storage for the default Jaeger
instance, the user has to create a Jaeger secret himself. As the
verrazzano-monitoring namespace will not be available prior to the
Verrazzano installation, the user is supposed to create the required
secrets in the verrazzano-install namespace. Changes are done similar to
fluentd external OpenSearch support to copy the secret from
verrazzano-install namespace to verrazzano-monitoring namespace, prior to
the creation of the Jaeger instance.