VZ-6686 Get the Service Account token from the secret created by VMC
Changes were done to create the Service Account token if there is
no secret associated with the Service Account. As the creation of
the secret with token is not reflected in the Service Account, the
sync_agent_secret code fails to get the token associated with the
Service Account. Changes have been made to get the secret token from the
secret created by the VMC for non-secret Service Accounts.