VZ-10091: Create Effective CR ConfigMap (#6327)
Added createOrUpdateEffectiveConfigCM() function which generates the Effective CR and stores it in a ConfigMap named <vz-name>-effective-config and is stored in the same namespace as verrazzano, i.e. <vz-namespace> as YAML. 
Added Code for deleting this ConfigMap upon VZ Uninstallation. 

Added Unit Tests for the function createOrUpdateEffectiveConfigCM()
Added Unit Tests for Deletion of ConfigMap upon VZ Uninstallation.